requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiohttp==3.11.2
python-dotenv~=1.0.1
pydantic==2.9.2
requests==2.32.3
pytest~=8.3.3
aioresponses~=0.7.7